+++ config/config.go
@@ -14,16 +14,6 @@
 
 package config
 
-import (
-	"errors"
-	"strings"
-
-	"github.com/go-playground/locales/en"
-	ut "github.com/go-playground/universal-translator"
-	"github.com/go-playground/validator/v10"
-	"go.uber.org/multierr"
-)
-
 const (
 	// ConfigKeyURL is the config name for a connection URL.
 	ConfigKeyURL = "url"
@@ -40,79 +30,6 @@ type Config struct {
 	// See https://www.postgresql.org/docs/current/sql-syntax-lexical.html#SQL-SYNTAX-IDENTIFIERS.
 	Table string `validate:"max=63"`
 	Key   string `validate:"max=63"`
-}
-
-// Validate validates the Config.
-func (c Config) Validate() error {
-	// init a translator and a universal translator
-	translator := en.New()
-	uni := ut.New(translator, translator)
-
-	uniTranslator, found := uni.GetTranslator("en")
-	if !found {
-		return errors.New("translator not found")
-	}
-
-	// init a new instance of a validator
-	validate := validator.New()
-
-	// register custom translations
-	if err := registerTranslations(validate, uniTranslator); err != nil {
-		return err
-	}
-
-	// collect all validation errors into one
-	if err := validate.Struct(c); err != nil {
-		var resultErr error
-		validationErrors := err.(validator.ValidationErrors)
-		for _, validationError := range validationErrors {
-			resultErr = multierr.Append(resultErr, errors.New(
-				validationError.Translate(uniTranslator),
-			))
-		}
-
-		return resultErr
-	}
-
-	return nil
-}
-
-// registerTranslations registers custom tranlations (error messages) for validation errors.
-func registerTranslations(validate *validator.Validate, uniTranslator ut.Translator) error {
-	// register a custom translation for the required tag
-	err := validate.RegisterTranslation("required", uniTranslator, func(ut ut.Translator) error {
-		return ut.Add("required", "\"{0}\" config value must be set", true)
-	}, func(ut ut.Translator, fe validator.FieldError) string {
-		t, _ := ut.T("required", fe.Field())
-		return strings.ToLower(t)
-	})
-	if err != nil {
-		return err
-	}
-
-	// register a custom translation for the url tag
-	err = validate.RegisterTranslation("url", uniTranslator, func(ut ut.Translator) error {
-		return ut.Add("url", "\"{0}\" config value must be a valid url", true)
-	}, func(ut ut.Translator, fe validator.FieldError) string {
-		t, _ := ut.T("url", fe.Field())
-		return strings.ToLower(t)
-	})
-	if err != nil {
-		return err
-	}
-
-	// register a custom translation for the max tag
-	err = validate.RegisterTranslation("max", uniTranslator, func(ut ut.Translator) error {
-		return ut.Add("max", "\"{0}\" config value is too long", true)
-	}, func(ut ut.Translator, fe validator.FieldError) string {
-		t, _ := ut.T("max", fe.Field())
-		return strings.ToLower(t)
-	})
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
 
 // Parse attempts to parse plugins.Config into a Config struct
